commit all test

+++ src/main/java/com/superhan/dvdrentalspringbootbackend/staff/repository/StaffRepository.java
@@ -3,5 +3,5 @@ package com.superhan.dvdrentalspringbootbackend.staff.repository;
 import com.superhan.dvdrentalspringbootbackend.staff.repository.entity.Staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StaffRepository extends JpaRepository<Staff, Long> {
+public interface StaffRepository extends JpaRepository<Staff, Integer> {
 }
